Location Revision
