+++ src/pages/home/home.js
@@ -4,7 +4,6 @@ import MapRenderer from '../../components/map/map';
 import Footer from '../../components/footer/footer'
 import Nav from '../../components/nav';
 import Banner from '../../components/banner/banner';
-import Location from '../..components/user-location/user-location';
 
 const Home = () => {
     const [results, setResults] = useState('');
@@ -20,6 +19,31 @@ const Home = () => {
             <Footer/>
         </main>
     )
+
+    
+      
+
+        if('geolocation' in navigator){
+            navigator.geolocation.getCurrentPosition(setPosition, showError);
+        }else{
+            
+            alert ("Browser doesn't Support Geolocation");
+        }
+        
+       /* function setPosition(position){
+            let latitude = position.coords.latitude;
+            let longitude = position.coords.longitude;
+            let local = [ latitude, longitude ];
+            console.log(latitude, longitude);
+    
+          window.localStorage.setItem("coordinates", JSON.stringify(local));
+        }; */
+        
+       function showError(error){
+            alert (` ${error.message}`) ;
+        }
+        
+            
 }
 
 export default Home;
